replaced card flip feature with a slide up bio
